requirements.txt
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
cryptography==41.0.1
django-cors-headers==4.1.0
requests==2.31.0
Pillow==10.0.0
ccxt==4.0.29
